// src/modules/school/teachers/teacher.interface.ts
export interface Teacher {
	first_name: string;
	last_name: string;
	gender: string;
	fathers_name: string;
	address: string;
	cnic: string;
	phone: string;
	dob: Date;
	qualification: string;
	yearOfGraduation: string;
	marksObtained: string;
	boardOrUniversity: string;
	designation: string;
	joining_date: Date;
	appointed_by: string;
	package: string;
}